fix db connection string, slight restructuring
