+++ backend/src/infra_service.rs
@@ -7,58 +7,7 @@ use phnxtypes::identifiers::Fqdn;
 use sqlx::{Executor, PgPool};
 use thiserror::Error;
 
-#[derive(Debug, Error)]
-pub enum StorageError {
-    #[error(transparent)]
-    Database(#[from] DatabaseError),
-    #[error("Error deserializing column: {0}")]
-    Serde(#[from] phnxtypes::codec::Error),
-}
-
-impl From<sqlx::Error> for StorageError {
-    fn from(e: sqlx::Error) -> Self {
-        Self::Database(e.into())
-    }
-}
-
-impl From<Box<dyn std::error::Error + Send + Sync>> for StorageError {
-    fn from(e: Box<dyn std::error::Error + Send + Sync>) -> Self {
-        Self::Database(e.into())
-    }
-}
-
-#[derive(Debug, Error)]
-pub enum DatabaseError {
-    #[error(transparent)]
-    Sqlx(#[from] sqlx::Error),
-    #[error(transparent)]
-    Dynamic(#[from] Box<dyn std::error::Error + Send + Sync>),
-}
-
-/// General error while accessing the requested queue.
-#[derive(Error, Debug)]
-pub(super) enum QueueError {
-    #[error(transparent)]
-    Storage(#[from] StorageError),
-    /// Mismatching sequence numbers.
-    #[error("Mismatching sequence numbers.")]
-    SequenceNumberMismatch,
-    /// Unrecoverable implementation error
-    #[error("Library Error")]
-    LibraryError,
-}
-
-impl From<sqlx::Error> for QueueError {
-    fn from(e: sqlx::Error) -> Self {
-        Self::Storage(e.into())
-    }
-}
-
-impl From<phnxtypes::codec::Error> for QueueError {
-    fn from(e: phnxtypes::codec::Error) -> Self {
-        Self::Storage(e.into())
-    }
-}
+use crate::{errors::StorageError, settings::DatabaseSettings};
 
 #[derive(Debug, Error)]
 pub enum ServiceCreationError {
@@ -77,12 +26,13 @@ impl<T: Into<sqlx::Error>> From<T> for ServiceCreationError {
 #[async_trait]
 pub trait InfraService: Sized {
     async fn new(
-        connection_string: &str,
-        db_name: &str,
+        database_settings: &DatabaseSettings,
         domain: Fqdn,
     ) -> Result<Self, ServiceCreationError> {
-        let connection = PgPool::connect(connection_string).await?;
+        let connection =
+            PgPool::connect(&database_settings.connection_string_without_database()).await?;
 
+        let db_name = database_settings.name.as_str();
         let db_exists = sqlx::query!(
             "select exists (
                 SELECT datname FROM pg_catalog.pg_database WHERE datname = $1
@@ -98,11 +48,9 @@ pub trait InfraService: Sized {
                 .await?;
         }
 
-        let connection_string_with_db = format!("{}/{}", connection_string, db_name);
         tracing::info!("Successfully created database {}", db_name);
-        tracing::info!("Connecting to postgres {}", connection_string_with_db);
 
-        let db_pool = PgPool::connect(&connection_string_with_db).await?;
+        let db_pool = PgPool::connect(&database_settings.connection_string()).await?;
 
         Self::new_from_pool(db_pool, domain).await
     }
