+++ backend/src/lib.rs
@@ -6,8 +6,10 @@
 
 pub mod auth_service;
 pub mod ds;
+pub mod errors;
+pub mod infra_service;
 pub mod messages;
-pub mod persistence;
 pub mod qs;
+pub mod settings;
 
 pub use mls_assist::messages::{AssistedGroupInfo, AssistedMessageOut};
