+++ server/src/main.rs
@@ -4,7 +4,7 @@
 
 use std::net::TcpListener;
 
-use phnxbackend::{auth_service::AuthService, ds::Ds, persistence::InfraService, qs::Qs};
+use phnxbackend::{auth_service::AuthService, ds::Ds, infra_service::InfraService, qs::Qs};
 use phnxserver::{
     configurations::*,
     endpoints::qs::{
@@ -53,12 +53,7 @@ async fn main() -> std::io::Result<()> {
         configuration.database.host
     );
     let mut counter = 0;
-    let mut ds_result = Ds::new(
-        &configuration.database.connection_string_without_database(),
-        &configuration.database.name,
-        domain.clone(),
-    )
-    .await;
+    let mut ds_result = Ds::new(&configuration.database, domain.clone()).await;
 
     // Try again for 10 times each second in case the postgres server is coming up.
     while let Err(e) = ds_result {
@@ -68,35 +63,22 @@ async fn main() -> std::io::Result<()> {
         if counter > 10 {
             panic!("Database not ready after 10 seconds.");
         }
-        ds_result = Ds::new(
-            &configuration.database.connection_string_without_database(),
-            &configuration.database.name,
-            domain.clone(),
-        )
-        .await;
+        ds_result = Ds::new(&configuration.database, domain.clone()).await;
     }
     let ds = ds_result.unwrap();
 
     // New database name for the QS provider
     configuration.database.name = format!("{}_qs", base_db_name);
     // QS storage provider
-    let qs = Qs::new(
-        &configuration.database.connection_string_without_database(),
-        &configuration.database.name,
-        domain.clone(),
-    )
-    .await
-    .expect("Failed to connect to database.");
+    let qs = Qs::new(&configuration.database, domain.clone())
+        .await
+        .expect("Failed to connect to database.");
 
     // New database name for the AS provider
     configuration.database.name = format!("{}_as", base_db_name);
-    let auth_service = AuthService::new(
-        &configuration.database.connection_string_without_database(),
-        &configuration.database.name,
-        domain.clone(),
-    )
-    .await
-    .expect("Failed to connect to database.");
+    let auth_service = AuthService::new(&configuration.database, domain.clone())
+        .await
+        .expect("Failed to connect to database.");
 
     let ws_dispatch_notifier = DispatchWebsocketNotifier::default_addr();
     let push_notification_provider = ProductionPushNotificationProvider::new(configuration.apns)
